+++ src/hooks/ready/checkCompendiums.js
@@ -36,8 +36,9 @@ export default async function () {
     let results = await Promise.allSettled([
       createIfNotExists("entity-spell-compendium", "Item", "Spells"),
       createIfNotExists("entity-item-compendium", "Item", "Items"),
-      createIfNotExists("entity-feature-compendium", "Item", "Features"),
+      createIfNotExists("entity-feature-compendium", "Item", "Class Features"),
       createIfNotExists("entity-class-compendium", "Item", "Classes"),
+      createIfNotExists("entity-trait-compendium", "Item", "Racial Traits"),
       createIfNotExists("entity-race-compendium", "Item", "Races"),
       createIfNotExists("entity-monster-compendium", "Actor", "Monsters"),
       // createIfNotExists("entity-monster-feature-compendium", "Item", "Monster Features")